// seaway.Web/src/app/Models/Enum/discount.ts
export enum DiscountRate{
  Below5Percent = 'Below 5%',
  FiveToTenPercent = '5% - 10%',
  ElevenToTwentyPercent = '11% - 20%',
  AboveTwentyPercent = 'Above 20%'
}

export enum PriceRange{
  Below5K = 'Below 5K',
  FiveKToTenK = '5K - 10K',
  Above10KTo20K = '11K - 20K',
  Above20KTo30K = '21K - 30K',
  Above30K = 'Above 30K'
}